+++ flutter_navigation_generator/lib/src/code_builder/import_builder.dart
@@ -20,11 +20,13 @@ class ImportBuilder {
     final imports = <String?>{
       'package:flutter/material.dart',
       'dart:convert',
-      if (routes.any((route) => route.guards?.isNotEmpty == true) || defaultGuards.isNotEmpty)
+      if (routes.any((route) => route.guards?.isNotEmpty == true) ||
+          defaultGuards.isNotEmpty)
         'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart',
     };
     imports.add(typeRefer(pageType, targetFile: targetFile).url);
-    imports.addAll(defaultGuards.map((e) => typeRefer(e, targetFile: targetFile).url));
+    imports.addAll(
+        defaultGuards.map((e) => typeRefer(e, targetFile: targetFile).url));
     imports.addAll(routes.expand(
       (route) => [
         typeRefer(route.routeWidget, targetFile: targetFile).url,
@@ -32,7 +34,8 @@ class ImportBuilder {
         typeRefer(route.returnType, targetFile: targetFile).url,
         ...route.parameters.expand((e) => [
               typeRefer(e, targetFile: targetFile).url,
-              ...e.typeArguments.map((e) => typeRefer(e, targetFile: targetFile).url),
+              ...e.typeArguments
+                  .map((e) => typeRefer(e, targetFile: targetFile).url),
             ]),
         ...?route.guards?.map((e) => typeRefer(e, targetFile: targetFile).url),
       ],