+++ flutter_navigation_generator/test/code_builder/on_generate_route_builder_test.dart
@@ -25,7 +25,8 @@ void main() {
       expect(content.requiredParameters.first.type?.symbol, 'RouteSettings');
       expect(content.requiredParameters.length, 1);
       expect(content.body is Block, true);
-      expect((content.body as Block).statements.first.toString(), 'return null;');
+      expect(
+          (content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -42,7 +43,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -77,7 +79,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -115,7 +118,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testWithoutReturn: return MaterialPageRoute<void>(builder: (_) => TestPageWithoutReturn(), settings: settings, fullscreenDialog: false,);case RouteNames.testWithReturn: return MaterialPageRoute<ReturnType>(builder: (_) => TestPageWithReturn(), settings: settings, fullscreenDialog: false,);}",
@@ -141,7 +145,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.r404: return MaterialPageRoute<void>(builder: (_) => Route404(), settings: settings, fullscreenDialog: false,);}",
@@ -168,7 +173,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage.constructorName(), settings: settings, fullscreenDialog: false,);}",
@@ -194,7 +200,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: true,);}",
@@ -223,7 +230,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -252,7 +260,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return CustomPageType<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);}",
@@ -274,7 +283,8 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+              const ImportableType(
+                  className: 'bool', name: 'testBool', isNullable: true),
               const ImportableType(className: 'int', name: 'testInt'),
             ],
           ),
@@ -282,7 +292,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -304,7 +315,8 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+              const ImportableType(
+                  className: 'bool', name: 'testBool', isNullable: true),
               const ImportableType(className: 'int', name: 'testInt'),
             ],
           ),
@@ -312,7 +324,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {}",
@@ -337,7 +350,8 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+              const ImportableType(
+                  className: 'bool', name: 'testBool', isNullable: true),
               const ImportableType(className: 'int', name: 'testInt'),
             ],
           ),
@@ -345,7 +359,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.testPage: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -367,7 +382,8 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+              const ImportableType(
+                  className: 'bool', name: 'testBool', isNullable: true),
               const ImportableType(className: 'int', name: 'testInt'),
             ],
           ),
@@ -375,7 +391,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {}",
@@ -399,7 +416,8 @@ void main() {
               className: 'TestPage',
             ),
             parameters: [
-              const ImportableType(className: 'bool', name: 'testBool', isNullable: true),
+              const ImportableType(
+                  className: 'bool', name: 'testBool', isNullable: true),
               const ImportableType(className: 'int', name: 'testInt'),
             ],
             defaultValues: {
@@ -411,7 +429,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: return MaterialPageRoute<void>(builder: (_) => TestPage(testBool: arguments['testBool'] is String ? arguments['testBool'] == 'true' : arguments['testBool'] as bool?,testInt: arguments['testInt'] is String ? int.parse(arguments['testInt']) : arguments['testInt'] as int,), settings: settings, fullscreenDialog: false,);}",
@@ -438,7 +457,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect((content.body as Block).statements.first.toString(), 'return null;');
+      expect(
+          (content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -456,7 +476,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect((content.body as Block).statements.first.toString(), 'return null;');
+      expect(
+          (content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -474,7 +495,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      expect((content.body as Block).statements.first.toString(), 'return null;');
+      expect(
+          (content.body as Block).statements.first.toString(), 'return null;');
       expect((content.body as Block).statements.length, 1);
     });
 
@@ -500,7 +522,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",
@@ -536,7 +559,8 @@ void main() {
       ).generate();
 
       expect(content.body is Block, true);
-      final bodyStatements = (content.body as Block).statements.map((f) => f.toString()).toList();
+      final bodyStatements =
+          (content.body as Block).statements.map((f) => f.toString()).toList();
       final expectedStatements = [
         "final arguments = settings.arguments is Map ? (settings.arguments as Map).cast<String, dynamic>() : <String, dynamic>{};\n    final settingsUri = Uri.parse(settings.name ?? '');\n    settingsUri.queryParameters.forEach((key, value) {\n      arguments[key] ??= value;\n    });",
         "switch (settingsUri.path) {case RouteNames.test: final testGuard = guards.whereType<TestGuard>().first;\nif (!testGuard.value) {\n  guardedRouteSettings = settings;\n  return onGenerateRoute(RouteSettings(\n    arguments: settings.arguments,\n    name: testGuard.alternativeRoute,\n  ));\n}\nreturn MaterialPageRoute<void>(builder: (_) => TestPage(), settings: settings, fullscreenDialog: false,);case RouteNames.test2: return MaterialPageRoute<void>(builder: (_) => TestPage2(), settings: settings, fullscreenDialog: false,);}",