+++ flutter_navigation_generator/test/utils/importable_type_to_string_test.dart
@@ -4,7 +4,9 @@ import 'package:flutter_navigation_generator/src/utils/importable_type_to_string
 import 'package:test/test.dart';
 
 void main() {
-  String convertType(String className) => ImportableTypeToStringUtils.convertFromString(ImportableType(className: className), 's');
+  String convertType(String className) =>
+      ImportableTypeToStringUtils.convertFromString(
+          ImportableType(className: className), 's');
   final mapRegex = RegExp(
       '[^<>]*<([^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>))?>))?>), ?([^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>)(?:, ?(?:[^<>]*|[^<>]*<(?:[^<>]*)(?:, ?(?:[^<>]*))?>))?>))?>))?>)>');
 
@@ -31,17 +33,30 @@ void main() {
 
   ImportableType? typeFromString(String string) {
     final match = mapRegex.firstMatch(string);
-    if (match == null) return ImportableType(className: string.replaceAll('?', ''), isNullable: string.endsWith('?'));
+    if (match == null) {
+      return ImportableType(
+          className: string.replaceAll('?', ''),
+          isNullable: string.endsWith('?'));
+    }
     final String? typeKey = match.group(1);
     final String? typeValue = match.group(2);
     if (typeKey == null) return ImportableType(className: string);
-    if (typeValue == null) return ImportableType(className: 'List', typeArguments: [typeFromString(typeKey)].whereNotNull().toList());
-    return ImportableType(className: 'Map', typeArguments: [typeFromString(typeKey), typeFromString(typeValue)].whereNotNull().toList());
+    if (typeValue == null) {
+      return ImportableType(
+          className: 'List',
+          typeArguments: [typeFromString(typeKey)].whereNotNull().toList());
+    }
+    return ImportableType(
+        className: 'Map',
+        typeArguments: [typeFromString(typeKey), typeFromString(typeValue)]
+            .whereNotNull()
+            .toList());
   }
 
   void checkMap(String type, String expected) {
     final importableType = typeFromString(type)!;
-    final result = ImportableTypeToStringUtils.convertFromString(importableType, 's');
+    final result =
+        ImportableTypeToStringUtils.convertFromString(importableType, 's');
     expect(result, expected);
   }
 
@@ -62,7 +77,8 @@ void main() {
 
     test('Custom object', () {
       const className = 'CustomObject';
-      const expected = 'CustomObject.fromJson(jsonDecode(utf8.decode(base64Decode(s))))';
+      const expected =
+          'CustomObject.fromJson(jsonDecode(utf8.decode(base64Decode(s))))';
       checkType(className, expected);
     });
   });
@@ -74,7 +90,11 @@ void main() {
           ? '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => e).toList()'
           : '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => e as $type).toList()';
       for (final testType in testTypes) {
-        checkImportableType(ImportableType(className: 'List', typeArguments: [ImportableType(className: testType)]), expectType(testType));
+        checkImportableType(
+            ImportableType(
+                className: 'List',
+                typeArguments: [ImportableType(className: testType)]),
+            expectType(testType));
       }
     });
 
@@ -84,30 +104,50 @@ void main() {
           ? '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => e).toList()'
           : '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => e as $type?).toList()';
       for (final testType in testTypes) {
-        checkImportableType(ImportableType(className: 'List', typeArguments: [ImportableType(className: testType, isNullable: true)]), expectType(testType));
+        checkImportableType(
+            ImportableType(className: 'List', typeArguments: [
+              ImportableType(className: testType, isNullable: true)
+            ]),
+            expectType(testType));
       }
     });
 
     test('List with custom type that might be nullable', () {
-      const expectNonNullable = '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => CustomObject.fromJson(e as Map<String, dynamic>)).toList()';
-      checkImportableType(const ImportableType(className: 'List', typeArguments: [ImportableType(className: 'CustomObject')]), expectNonNullable);
+      const expectNonNullable =
+          '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => CustomObject.fromJson(e as Map<String, dynamic>)).toList()';
+      checkImportableType(
+          const ImportableType(
+              className: 'List',
+              typeArguments: [ImportableType(className: 'CustomObject')]),
+          expectNonNullable);
 
-      const expectNullable = '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => CustomObject?.fromJson(e as Map<String, dynamic>)).toList()';
-      checkImportableType(const ImportableType(className: 'List', typeArguments: [ImportableType(className: 'CustomObject', isNullable: true)]), expectNullable);
+      const expectNullable =
+          '(jsonDecode(utf8.decode(base64Decode(s))) as List<dynamic>).map((e) => e == null ? null : CustomObject.fromJson(e as Map<String, dynamic>)).toList()';
+      checkImportableType(
+          const ImportableType(className: 'List', typeArguments: [
+            ImportableType(className: 'CustomObject', isNullable: true)
+          ]),
+          expectNullable);
     });
   });
 
   group('Maps', () {
     test('Map with basic types', () {
-      checkMap('Map<String, dynamic>', 'Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
-      checkMap('Map<String, int>', 'Map<String, int>.from(jsonDecode(utf8.decode(base64Decode(s))))');
-      checkMap('Map<int, int>', 'Map<int, int>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<String, dynamic>',
+          'Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<String, int>',
+          'Map<String, int>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<int, int>',
+          'Map<int, int>.from(jsonDecode(utf8.decode(base64Decode(s))))');
     });
 
     test('Map with basic types nullable', () {
-      checkMap('Map<String?, dynamic>', 'Map<String?, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
-      checkMap('Map<String, int?>', 'Map<String, int?>.from(jsonDecode(utf8.decode(base64Decode(s))))');
-      checkMap('Map<int?, int?>', 'Map<int?, int?>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<String?, dynamic>',
+          'Map<String?, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<String, int?>',
+          'Map<String, int?>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('Map<int?, int?>',
+          'Map<int?, int?>.from(jsonDecode(utf8.decode(base64Decode(s))))');
     });
 
     test('Map with custom types', () {
@@ -144,16 +184,20 @@ void main() {
           'Map<String, Map>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(k, Map<String, Map>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(k, Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(v))))))))))');
       checkMap('Map<Map<Map<String, dynamic>, dynamic>, String>',
           'Map<Map, String>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(Map<Map, dynamic>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(k)))), v))), v)))');
-      checkMap('Map<Map<Map<String, dynamic>, Map<String, dynamic>>, Map<Map<String, dynamic>, Map<String, dynamic>>>',
+      checkMap(
+          'Map<Map<Map<String, dynamic>, Map<String, dynamic>>, Map<Map<String, dynamic>, Map<String, dynamic>>>',
           'Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(k)))), Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(v))))))), Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(k)))), Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(v))))))))))');
     });
 
     test('Map with map with map with custom types', () {
-      checkMap('Map<CustomObject, Map<CustomObject, Map<CustomObject, dynamic>>>',
+      checkMap(
+          'Map<CustomObject, Map<CustomObject, Map<CustomObject, dynamic>>>',
           'Map<CustomObject, Map>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), Map<CustomObject, Map>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), Map<CustomObject, dynamic>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), v)))))))))');
-      checkMap('Map<Map<Map<String, CustomObject>, CustomObject>, CustomObject>',
+      checkMap(
+          'Map<Map<Map<String, CustomObject>, CustomObject>, CustomObject>',
           'Map<Map, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(Map<Map, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(Map<String, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(k, CustomObject.fromJson(v as Map<String, dynamic>)))), CustomObject.fromJson(v as Map<String, dynamic>)))), CustomObject.fromJson(v as Map<String, dynamic>))))');
-      checkMap('Map<Map<Map<String, CustomObject>, Map<String, dynamic>>, Map<Map<CustomObject, dynamic>, Map<String, CustomObject>>>',
+      checkMap(
+          'Map<Map<Map<String, CustomObject>, Map<String, dynamic>>, Map<Map<CustomObject, dynamic>, Map<String, CustomObject>>>',
           'Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(Map<String, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(k, CustomObject.fromJson(v as Map<String, dynamic>)))), Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(v))))))), Map<Map, Map>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(Map<CustomObject, dynamic>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), v))), Map<String, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(k, CustomObject.fromJson(v as Map<String, dynamic>))))))))))');
     });
 
@@ -167,18 +211,21 @@ void main() {
     });
 
     test('List with Map', () {
-      checkMap('List<Map<String, dynamic>>', 'Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
+      checkMap('List<Map<String, dynamic>>',
+          'Map<String, dynamic>.from(jsonDecode(utf8.decode(base64Decode(s))))');
       checkMap('List<Map<CustomObject, CustomObject>>',
           'Map<CustomObject, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), CustomObject.fromJson(v as Map<String, dynamic>))))');
     });
 
     test('Let\'s go deeper', () {
-      checkMap('List<Map<double, List<Map<List<Map<CustomObject, CustomObject>>, int>>>>',
+      checkMap(
+          'List<Map<double, List<Map<List<Map<CustomObject, CustomObject>>, int>>>>',
           'Map<double, Map>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(k, Map<Map, int>.from(jsonDecode(utf8.decode(base64Decode(v))).map((k, v) => MapEntry(Map<CustomObject, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(CustomObject.fromJson(k as Map<String, dynamic>), CustomObject.fromJson(v as Map<String, dynamic>)))), v))))))');
     });
 
     test('Let\'s go deeper with nullables', () {
-      checkMap('List<Map<double?, List<Map<List<Map<CustomObject?, CustomObject>>, int>>?>>',
+      checkMap(
+          'List<Map<double?, List<Map<List<Map<CustomObject?, CustomObject>>, int>>?>>',
           'Map<Map, int>.from(jsonDecode(utf8.decode(base64Decode(s))).map((k, v) => MapEntry(Map<CustomObject?, CustomObject>.from(jsonDecode(utf8.decode(base64Decode(k))).map((k, v) => MapEntry(k == null ? null : CustomObject.fromJson(k as Map<String, dynamic>), CustomObject.fromJson(v as Map<String, dynamic>)))), v)))');
     });
   });