+++ flutter_navigation_generator/CHANGELOG.md
@@ -1,9 +1,11 @@
-## 1.1.0-experimental (2024-06-25)
+## 2.0.0 (2024-07-21)
 
 - Added support for web parameters as query parameters and url path parameters
 - Added support for methodName
 - Added a package (flutter_navigation_generator_animations) with NativeRouteAnimation to animate between pages in a native way (Android scales, iOS uses CupertinoPageTransition, web doesn't animate, other platforms use FadeTransition)
+- Added unknownRoute to handle unknown routes
 - Added guards to prevent navigation under certain conditions
+- Added defaultGuards to apply to all routes where no guards are specified
 
 ## 1.0.4+1 (2024-03-06)