+++ flutter_navigation_generator_animations/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 1.1.0-experimental (2024-07-03)
+## 2.0.0 (2024-07-21)
 
 - Added NativeRouteAnimation to animate between pages in a native way (Android scales, iOS uses CupertinoPageTransition, web doesn't animate, other platforms use FadeTransition)