+++ flutter_navigation_generator_annotations/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.1.0-experimental (2024-07-03)
+## 2.0.0 (2024-07-21)
 
 - Added methodName and updated routeName to include web parameters
 - Added unknownRoute to handle unknown routes